# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web Scraper Public API",
    "version": "1.0.0",
    "description": "Authenticate with `Authorization: Bearer <API_KEY>`. Create keys in the dashboard under API keys. Jobs are asynchronous: create, poll `GET /api/v1/jobs/{id}`, then read results. Errors always have the shape `{ error: { code, message, details?, requestId } }`."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wsk_..."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "summary": "List jobs",
        "tags": [
          "jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "project_id",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "queued"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "running"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "completed"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "failed"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "cancelled"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "status",
            "required": false,
            "description": "queued -> running -> completed | failed | cancelled"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "scrape"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "crawl"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "before",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "summary": "Get job status",
        "tags": [
          "jobs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "project_id",
                    "type",
                    "target_url",
                    "status",
                    "source",
                    "options",
                    "progress",
                    "error",
                    "created_at",
                    "started_at",
                    "completed_at",
                    "duration_ms",
                    "links"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "project_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "scrape"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "crawl"
                          ]
                        }
                      ]
                    },
                    "target_url": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "queued"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "running"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "completed"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "failed"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "cancelled"
                          ]
                        }
                      ],
                      "description": "queued -> running -> completed | failed | cancelled"
                    },
                    "source": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "dashboard"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "api"
                          ]
                        }
                      ]
                    },
                    "options": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "progress": {
                      "type": "object",
                      "required": [
                        "pages_discovered",
                        "pages_processed",
                        "pages_succeeded",
                        "pages_failed"
                      ],
                      "properties": {
                        "pages_discovered": {
                          "type": "integer"
                        },
                        "pages_processed": {
                          "type": "integer"
                        },
                        "pages_succeeded": {
                          "type": "integer"
                        },
                        "pages_failed": {
                          "type": "integer"
                        }
                      }
                    },
                    "error": {
                      "anyOf": [
                        {
                          "type": "object",
                          "required": [
                            "code",
                            "message"
                          ],
                          "properties": {
                            "code": {
                              "type": "string"
                            },
                            "message": {
                              "type": "string"
                            }
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "started_at": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "completed_at": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "duration_ms": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "links": {
                      "type": "object",
                      "required": [
                        "self",
                        "results"
                      ],
                      "properties": {
                        "self": {
                          "type": "string"
                        },
                        "results": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "title": "Job"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable error code",
                          "example": "INVALID_URL"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "title": "Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}